Modifies 0-print_listint.c

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -2,12 +2,13 @@
 
 /**
  * print_listint - Prints all the elements of a listint_t list
- * @h: A pointer to listint_t
+ * @h: A pointer to the start of the list
+ *
  * Return: the number of nodes
 */
 size_t print_listint(const listint_t *h)
 {
-	unsigned int nodes = 0;
+	size_t nodes = 0;
 
 	while (h != NULL)
 	{
